modify_image/requirements.txt
iniconfig==2.3.0
numpy==2.3.5
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.2
